+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to 
 [Semantic Versioning](http://semver.org/).
 
-## 1.31.0.dev (development stage/unreleased/unstable)
+## 1.32.0.dev (development stage/unreleased/unstable)
+
+## 1.32.0
 Now `stream_buffer` can be used as FIFO or LIFO stack and its possible to define a max length for it.
 ### Added
 - `clear_stream_buffer()` to delete all items on the `stream_buffer` stack.